convert our uint64's to use LEB128 encoding (#598)
Fixes #587

In which we convert the uint64's to a new type, Uint64 which we teach refmt to serialize in LEB128. Why you may ask does this PR marshal Uint64 to json as a string instead of a number? See #599

+++ actor/builtin/paymentbroker/paymentbroker.go
@@ -160,7 +160,7 @@ func (pb *Actor) CreateChannel(ctx *vm.Context, target types.Address, eol *types
 			storage.Channels[ctx.Message().From.String()] = byPayer
 		}
 
-		channelID := types.NewChannelID(ctx.Message().Nonce)
+		channelID := types.NewChannelID(uint64(ctx.Message().Nonce))
 
 		if _, found := byPayer[channelID.String()]; found {
 			return nil, Errors[ErrDuplicateChannel]
